refactor: group scipy compat
